@@ -6,38 +6,36 @@
     "start": "ng serve",
     "build": "ng build",
     "build:test": "ng build --configuration=test",
-    "build:prod": "ng build --configuration=production",
+    "build:prod": "ng build --configuration=production --aot",
     "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.2.14",
-    "@angular/cdk": "^8.2.3",
-    "@angular/common": "^8.2.14",
-    "@angular/compiler": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/forms": "^8.2.14",
-    "@angular/localize": "^9.1.13",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
-    "@angular/router": "^8.2.14",
-    "@auth0/angular-jwt": "^4.0.0",
-    "@fortawesome/angular-fontawesome": "^0.3.0",
+    "@angular/animations": "^17.3.1",
+    "@angular/cdk": "^17.3.1",
+    "@angular/common": "^17.3.1",
+    "@angular/compiler": "^17.3.1",
+    "@angular/core": "^17.3.1",
+    "@angular/forms": "^17.3.1",
+    "@angular/localize": "^17.3.1",
+    "@angular/platform-browser": "^17.3.1",
+    "@angular/platform-browser-dynamic": "^17.3.1",
+    "@angular/router": "^17.3.1",
+    "@auth0/angular-jwt": "^5.2.0",
+    "@fortawesome/angular-fontawesome": "^0.14.0",
     "@fortawesome/fontawesome-free": "^5.15.3",
     "@fortawesome/fontawesome-svg-core": "^1.2.35",
-    "@fortawesome/free-regular-svg-icons": "^5.15.3",
     "@fortawesome/free-solid-svg-icons": "^5.15.3",
     "@types/chart.js": "^2.9.3",
     "@types/heatmap.js": "^2.0.36",
-    "@types/leaflet": "^1.7.0",
-    "@types/leaflet.heat": "^0.2.0",
     "animate.css": "^4.1.0",
-    "chart.js": "^2.9.4",
-    "chartjs-plugin-datalabels": "^1.0.0",
-    "core-js": "^2.5.4",
+    "bootstrap": "^5.3.3",
+    "chart.js": "^4.4.2",
+    "chartjs-plugin-datalabels": "^2.2.0",
     "heatmap.js": "git+https://github.com/ionata/heatmap.js.git",
+    "html2canvas": "^1.4.1",
     "jwt-decode": "^2.2.0",
     "leaflet": "^1.7.1",
     "leaflet-heatmap": "^1.0.0",
@@ -45,37 +43,38 @@
     "lodash": "^4.17.21",
     "moment": "^2.27.0",
     "moment-timezone": "^0.5.33",
-    "ng-recaptcha": "^5.0.0",
+    "ng-recaptcha": "^13.2.1",
     "primeflex": "^1.0.0",
     "primeicons": "^2.0.0",
-    "primeng": "^8.1.1",
-    "prismjs": "^1.24.1",
-    "rxjs": "~6.3.3",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.8.26",
-    "bootstrap": "^5.3.3"
+    "primeng": "^17.3.1",
+    "qs": "^6.12.0",
+    "rxjs": "~6.6.7",
+    "source-map-support": "^0.5.21",
+    "zone.js": "~0.14.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.29",
-    "@angular/cli": "^8.3.29",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/language-service": "^8.2.14",
-    "@types/geojson": "7946.0.8",
-    "@types/jasmine": "^2.8.18",
+    "@angular-devkit/build-angular": "^17.3.1",
+    "@angular/cli": "^17.3.1",
+    "@angular/compiler-cli": "^17.3.1",
+    "@angular/language-service": "^17.3.1",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "^2.0.10",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
+    "@types/node": "^12.11.1",
+    "@typescript-eslint/eslint-plugin": "^7.4.0",
+    "@typescript-eslint/parser": "^7.4.0",
+    "codelyzer": "^6.0.0",
+    "eslint": "^8.57.0",
+    "jasmine-core": "~5.1.2",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~6.4.3",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "^2.1.0",
+    "protractor": "~7.0.0",
     "sass": "^1.49.5",
     "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.5.3"
+    "tslint": "~6.1.0",
+    "typescript": "~5.4.3"
   }
 }
