@@ -2,7 +2,7 @@ import * as _ from 'lodash';
 
 import {Component, OnInit} from '@angular/core';
 import {MessageService, SelectItem} from 'primeng/api';
-import {FormBuilder, FormGroup, Validators} from '@angular/forms';
+import {UntypedFormBuilder, UntypedFormGroup, Validators} from '@angular/forms';
 
 import {AuthService} from '../../shared/services/auth.service';
 import {ConferenceService} from '../../shared/services/conference.service';
@@ -20,13 +20,13 @@ import {StoreKeys} from '../../shared/commons/contants';
 })
 export class CompleteProfileComponent implements OnInit {
 
-  userForm: FormGroup;
+  userForm: UntypedFormGroup;
   localities: ILocality[] = [];
   filteredLocalities: SelectItem[];
   localityType: string;
 
   constructor(
-    private formBuilder: FormBuilder,
+    private formBuilder: UntypedFormBuilder,
     private localitySrv: LocalityService,
     private authSrv: AuthService,
     private personSrv: PersonService,
@@ -57,27 +57,30 @@ export class CompleteProfileComponent implements OnInit {
       contactEmail: [_.get(value, 'contactEmail', ''), Validators.required],
       telephone: [_.get(value, 'telephone', '')],
       locality: [_.get(value, 'locality', ''), Validators.required],
-      receiveInformational: false,
+      receiveInformational: true,
     });
   }
 
 
   async save({id, name, contactEmail, telephone, locality, receiveInformational}) {
     const sender: IPerson = {
-      id, name, contactEmail, telephone,
+      id, name, contactEmail, telephone, receiveInformational,
       selfDeclaration: {
         conference: this.conferenceSrv.ConferenceActiveId,
-        locality: _.get(locality, 'id'),
-        receiveInformational,
+        locality: locality,
       }
     };
 
     const {success} = await this.personSrv.complementPerson(sender);
     if (success) {
 
-      if(localStorage.getItem(StoreKeys.CHECK_IN)){
+      if(sessionStorage.getItem(StoreKeys.CHECK_IN)){
         localStorage.removeItem(StoreKeys.IS_PROFILE_INCOMPLETED);
-        await this.router.navigate(['/self-check-in']);
+        await this.router.navigate([`/self-check-in/${localStorage.getItem(StoreKeys.CONFERENCE_ACTIVE)}/meeting/${sessionStorage.getItem(StoreKeys.CHECK_IN)}`]);
+      }else if(sessionStorage.getItem(StoreKeys.PRE_REGISTRATION)){
+        localStorage.removeItem(StoreKeys.IS_PROFILE_INCOMPLETED);
+        localStorage.removeItem(StoreKeys.REDIRECT_URL)
+        await this.router.navigate([`/registration/${localStorage.getItem(StoreKeys.CONFERENCE_ACTIVE)}/meeting/${sessionStorage.getItem(StoreKeys.PRE_REGISTRATION)}`]);
       }else{
         localStorage.removeItem(StoreKeys.IS_PROFILE_INCOMPLETED);
         await this.router.navigate(['/conference-map']);
