@@ -57,11 +57,11 @@ import { StartComponent } from './pages/start/start.component';
 import { PreRegistrationComponent } from './pages/pre-registration/pre-registration.component';
 import { registerLocaleData } from '@angular/common';
 import localePtBr from '@angular/common/locales/pt';
-import { TriStateCheckboxModule } from 'primeng/components/tristatecheckbox/tristatecheckbox';
+import { TriStateCheckboxModule } from 'primeng/tristatecheckbox';
 import { SafeHTMLPipe } from './shared/pipes/safeHTML/safe-html.pipe';
-import { ButtonModule } from 'primeng/components/button/button';
-import { LoginSelfCheckInComponent } from './pages/login-self-check-in/login-self-check-in.component';
+import { ButtonModule } from 'primeng/button';
 import { SelfCheckInComponent } from './pages/self-check-in/self-check-in.component';
+import { LoginPreRegistrationSelfCheckInComponent } from './pages/login-pre-registration-self-check-in/login-pre-registration-self-check-in.component';
 
 
 registerLocaleData(localePtBr);
@@ -69,8 +69,8 @@ registerLocaleData(localePtBr);
 @NgModule({
   declarations: [
     AppComponent,
-    LoginSelfCheckInComponent,
     StartComponent,
+    LoginPreRegistrationSelfCheckInComponent,
     LoginComponent,
     CompleteProfileComponent,
     TemplateComponent,
