@@ -49,8 +49,9 @@ export class ConferenceMapComponent implements OnInit {
         if (this.regionalization) {
           const { success, data } = await this.localitySrv.getConferenceCards(this.conferenceSrv.ConferenceActiveId);
           if (success) {
-            let localidadesExibicao = ['Central Sul','Caparaó','Nordeste','Rio Doce','Noroeste', 'Centro Oeste', 'Sudoeste Serrana', 'Central Serrana','Metropolitana','Litoral Sul'];
-            data.localities = data.localities.filter(obj => localidadesExibicao.filter(n => n === obj.name).length > 0 );
+            // Utilizado no ES500
+            // const localidadesExibicao = ['Central Sul','Caparaó',];
+            // data.localities = data.localities.filter(obj => localidadesExibicao.filter(n => n === obj.name).length > 0 );
             this.conference = data;
             this.breadcrumbSrv.add({ title: data.regionalizable});
           }
