@@ -12,6 +12,7 @@ import {INavigation} from '../../shared/interfaces/INavigation';
 import * as _ from 'lodash';
 import {MessageService} from 'primeng/api';
 import {animate, state, style, transition, trigger} from '@angular/animations';
+import { ColorService } from 'src/app/shared/services/color.service';
 
 @Component({
   selector: 'app-conference-steps',
@@ -55,7 +56,7 @@ export class ConferenceStepsComponent implements OnInit, OnDestroy {
   hide: string = 'default';
   textSearch: string = '';
   state: string = 'default';
-  imageName: string = 'search';
+  imageName: string = 'search_svg';
   wasFound: boolean = false;
   searchMessage: string;
   disable: boolean = false;
@@ -69,7 +70,8 @@ export class ConferenceStepsComponent implements OnInit, OnDestroy {
     private route: Router,
     private activeRoute: ActivatedRoute,
     private participationStateSrv: ParticipationStateService,
-    private messageService: MessageService
+    private messageService: MessageService,
+    protected colorService: ColorService,
   ) {
   }
 
@@ -107,22 +109,35 @@ export class ConferenceStepsComponent implements OnInit, OnDestroy {
 
   async populateState() {
     const participationState = this.participationStateSrv.getSync();
-    this.localityId = _.get(participationState, 'locality.id');
+    this.localityId = _.get(participationState, 'locality.id') as number;
     this.navigation = _.get(participationState, 'navigation', []);
   }
 
   async loadConferenceStep() {
-    const {success, data} = await this.participationSrv.getPlanItem(
-//      this.conferenceSrv.ConferenceActiveId, this.localityId, this.stepId ? this.stepId : null, ''
-      this.conference.id, this.localityId, this.stepId ? this.stepId : null, ''
-    );
-    if (success) {
-      this.conferenceStepItem = data;
-      this.conferenceStepItem.itens.forEach(item => {
-        item.checked = (item.votes || item.commentsMade > 0);
-      });
-      this.breadcrumbSrv.addOrUpdate({title: data.structureitem.name});
-      this.participationStateSrv.removeBeforeNavigation(data.structureitem.name);
+
+    if (this.stepId) {
+      const { success, data } = await this.participationSrv.getPlanItemChildren(
+        this.conference.id, this.localityId, this.stepId);
+      if (success) {
+        this.conferenceStepItem = data;
+        this.conferenceStepItem.itens.forEach(item => {
+          item.checked = (item.votes || item.commentsMade > 0);
+        });
+        this.breadcrumbSrv.addOrUpdate({ title: data.structureitem.name });
+        this.participationStateSrv.removeBeforeNavigation(data.structureitem.name);
+      }
+    } else {
+      const { success, data } = await this.participationSrv.getPlanItem(
+        this.conference.id, this.localityId, null, ''
+      );
+      if (success) {
+        this.conferenceStepItem = data;
+        this.conferenceStepItem.itens.forEach(item => {
+          item.checked = (item.votes || item.commentsMade > 0);
+        });
+        this.breadcrumbSrv.addOrUpdate({ title: data.structureitem.name });
+        this.participationStateSrv.removeBeforeNavigation(data.structureitem.name);
+      }
     }
   }
 
@@ -264,22 +279,19 @@ export class ConferenceStepsComponent implements OnInit, OnDestroy {
     this.state = (this.state === 'default' ? 'rotated' : 'default');
     if (this.hide === 'default') {
       await this.delay(600);
-      this.imageName = 'search';
+      this.imageName = 'search_svg';
       this.disable = false;
     } else {
-      this.imageName = 'close';
+      this.imageName = 'close_svg';
     }
-    const {success, data} = await this.participationSrv
-      .getPlanItem(this.conferenceSrv.ConferenceActiveId, this.localityId, this.stepId, this.textSearch);
 
-    if (success) {
-      this.renderList(data.itens);
-    }
   }
 
   async search() {
-    const {success, data} = await this.participationSrv
-      .getPlanItem(this.conferenceSrv.ConferenceActiveId, this.localityId, this.stepId, this.textSearch);
+
+    if(this.stepId){
+      const {success, data} = await this.participationSrv
+      .getPlanItemChildren(this.conferenceSrv.ConferenceActiveId, this.localityId, this.stepId, this.textSearch);
 
     if (success) {
       if (!data.itens || data.itens.length < 1) {
@@ -295,6 +307,26 @@ export class ConferenceStepsComponent implements OnInit, OnDestroy {
       this.showMessage = true;
       this.classMsg = 'animate__animated animate__fadeInRightBig';
     }
+    }else{
+      const {success, data} = await this.participationSrv
+        .getPlanItem(this.conferenceSrv.ConferenceActiveId, this.localityId, this.stepId, this.textSearch);
+  
+      if (success) {
+        if (!data.itens || data.itens.length < 1) {
+          this.searchMessage = 'Hummm... Não estou encontrando esse termo. Que tal tentar um sinônimo ou algo menos específico?';
+          this.wasFound = false;
+          this.renderList([]);
+        } else {
+          this.searchMessage = 'Oba! Encontrei alguma coisa nos itens abaixo!';
+          this.wasFound = true;
+          this.renderList(data.itens);
+        }
+  
+        this.showMessage = true;
+        this.classMsg = 'animate__animated animate__fadeInRightBig';
+      }
+    }
+    
   }
 
   renderList(data: IItem[]) {
