@@ -5,7 +5,10 @@ import { AuthService } from "src/app/shared/services/auth.service";
 import { MeetingService } from "src/app/shared/services/meeting.service";
 import 'moment-timezone';
 import { FormBuilder, FormGroup, Validators } from "@angular/forms";
-import { Router } from "@angular/router";
+import { ActivatedRoute, Router } from "@angular/router";
+import { ConferenceService } from "src/app/shared/services/conference.service";
+import { IConference } from "src/app/shared/interfaces/IConference";
+import { ColorService } from "src/app/shared/services/color.service";
 
 
 @Component({
@@ -19,22 +22,45 @@ export class SelfCheckInComponent implements OnInit {
     loading = true;
     checkInData = [{}]
     isNotEmptyCheckInObject = false
+    conference: IConference
 
     constructor(
+      private conferenceSrv: ConferenceService,
       private formBuilder: FormBuilder,
       private authSrv: AuthService,
-      protected meetingSrv: MeetingService,
-      private router: Router
+      private meetingSrv: MeetingService,
+      private router: Router,
+      private route: ActivatedRoute,
+      private colorService: ColorService
     ){}
 
     
-    ngOnInit() {
-      this.registerAttendance(this.authSrv.getUserInfo.id, JSON.parse(sessionStorage.getItem(StoreKeys.CHECK_IN)))
+    async ngOnInit() {
+      const selfcheckInIsOpen = await this.meetingSrv.getSelfCheckInOrPreRegistrationOpen(this.route.snapshot.params['meeting'],"self-check-in")
+      if(selfcheckInIsOpen.data.length == 0){
+        await sessionStorage.setItem(StoreKeys.CHECK_IN_OFF, this.route.snapshot.params['meeting']);
+        await localStorage.setItem(StoreKeys.CONFERENCE_ACTIVE,this.route.snapshot.params['conference']);
+        await this.colorService.setPrimaryColor(localStorage.getItem(StoreKeys.CONFERENCE_ACTIVE))
+        this.router.navigate(['/login-pre-registration-self-check-in']);
+      }else if(!sessionStorage.getItem(StoreKeys.LOGIN_CHECK_IN)){
+        await this.loadingVariable()
+        await this.colorService.setPrimaryColor(localStorage.getItem(StoreKeys.CONFERENCE_ACTIVE))
+        this.router.navigate(['/login-pre-registration-self-check-in']);
+      }else{
+        sessionStorage.removeItem(StoreKeys.CHECK_IN)
+        this.registerAttendance(this.authSrv.getUserInfo.id, JSON.parse(this.route.snapshot.params['meeting']))
+      }
+    }
+
+    async loadingVariable(){
+      await sessionStorage.setItem(StoreKeys.CHECK_IN, this.route.snapshot.params['meeting']);
+      await sessionStorage.setItem(StoreKeys.LOGIN_CHECK_IN, 'true');
+      await localStorage.setItem(StoreKeys.CONFERENCE_ACTIVE,this.route.snapshot.params['conference']);
     }
 
     async isCheckin(meeting, personId){
-      var now = new Date();
-      var timeZone = now.toString().split(' ')[5];
+      const now = new Date();
+      const timeZone = now.toString().split(' ')[5];
      await this.meetingSrv.postCheckIn(meeting, personId, timeZone)
     }
 
@@ -48,11 +74,15 @@ export class SelfCheckInComponent implements OnInit {
     }
 
     async registerAttendance(personId, meeting){
+      await this.conferenceSrv.GetById(parseInt(this.route.snapshot.params['conference'])).then(
+        (res) => {
+          this.conference = res.data
+      })
       await this.getPersonAndMeeting(personId, meeting)
       if(this.isNotEmptyCheckInObject){
         this.setForm(this.checkInData)
       }else{
-        await this.isCheckin(sessionStorage.getItem(StoreKeys.CHECK_IN), this.authSrv.getUserInfo.id)
+        await this.isCheckin(meeting, this.authSrv.getUserInfo.id)
         await this.getPersonAndMeeting(personId, meeting)
         this.setForm(this.checkInData)
       }
@@ -64,12 +94,13 @@ export class SelfCheckInComponent implements OnInit {
         personName: [_.get(value.data.person, 'name', ''), Validators.required],
         date: [_.get(value.data, 'time', ''), Validators.required],
         meetingName: [_.get(value.data.meeting, 'name', ''), Validators.required],
-        localityPlace: [_.get(value.data.meeting.localityPlace, 'name', ''), Validators.required]
+        localityPlace: [_.get(value.data.meeting.localityPlace, 'name', ''), Validators.required],
+        conferenceName:[_.get(this.conference, 'name', ''), Validators.required]
       });
     }
 
     cancel(){
-      sessionStorage.removeItem(StoreKeys.CHECK_IN);
+      sessionStorage.removeItem(StoreKeys.LOGIN_CHECK_IN)
       this.router.navigate(['/conference-map']);
   }
   }
